requirements.txt
pillow
sentence_transformers
chromadb
langchain
langchain_community
Flask
flask-limiter
flask-cors
open-clip-torch